requirements/requirements-dev.txt
#
# This file is autogenerated by pip-compile with Python 3.12
# by the following command:
#
#    pip-compile --extra=dev --output-file=requirements/requirements-dev.txt pyproject.toml
#
build==1.2.1
    # via pip-tools
click==8.1.7
    # via pip-tools
iniconfig==2.0.0
    # via pytest
numpy==2.1.0
    # via
    #   pandas
    #   pyarrow
packaging==24.1
    # via
    #   build
    #   pytest
pandas==2.2.2
    # via hub-dashboard-predtimechart (pyproject.toml)
pip-tools==7.4.1
    # via hub-dashboard-predtimechart (pyproject.toml)
pluggy==1.5.0
    # via pytest
pyarrow==17.0.0
    # via hub-dashboard-predtimechart (pyproject.toml)
pyproject-hooks==1.1.0
    # via
    #   build
    #   pip-tools
pytest==8.3.2
    # via hub-dashboard-predtimechart (pyproject.toml)
python-dateutil==2.9.0.post0
    # via pandas
pytz==2024.1
    # via pandas
pyyaml==6.0.2
    # via hub-dashboard-predtimechart (pyproject.toml)
six==1.16.0
    # via python-dateutil
structlog==24.4.0
    # via hub-dashboard-predtimechart (pyproject.toml)
tzdata==2024.1
    # via pandas
wheel==0.44.0
    # via pip-tools

# The following packages are considered to be unsafe in a requirements file:
# pip
# setuptools
